Factor out container deployer.
